backend/requirements.txt
# NexusAI - Multimodal Agentic Assistant
# Core Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.1

# LLM Interface
ollama==0.4.7

# RAG / Vector Memory
chromadb==0.6.3
sentence-transformers==3.3.1

# Speech-to-Text
openai-whisper==20240930

# Document Processing
PyPDF2==3.0.1
Pillow==11.1.0

# Utilities
pydantic==2.10.5
python-dateutil==2.9.0
